+++ apps/website/src/components/layout/header.tsx
@@ -1,15 +1,9 @@
 import * as React from 'react'
 
-import {
-  Box,
-  BoxProps,
-  Container,
-  HStack,
-  useColorModeValue,
-} from '@chakra-ui/react'
+import { Box, BoxProps, Container, HStack } from '@chakra-ui/react'
 import Navigation from './navigation'
 import Logo from './logo'
-import { useViewportScroll } from 'framer-motion'
+import { useScroll } from 'framer-motion'
 
 export interface HeaderProps extends Omit<BoxProps, 'children'> {}
 
@@ -18,12 +12,10 @@ const Header = (props: HeaderProps) => {
   const [y, setY] = React.useState(0)
   const { height = 0 } = ref.current?.getBoundingClientRect() ?? {}
 
-  const { scrollY } = useViewportScroll()
+  const { scrollY } = useScroll()
   React.useEffect(() => {
     return scrollY.onChange(() => setY(scrollY.get()))
   }, [scrollY])
-
-  const bg = useColorModeValue('whiteAlpha.700', 'hsl(210 11% 10% / 85%)')
 
   return (
     <Box
@@ -31,33 +23,45 @@ const Header = (props: HeaderProps) => {
       as="header"
       top="0"
       w="full"
-      backdropFilter="blur(16px)"
       zIndex="sticky"
       borderColor="whiteAlpha.100"
       transitionProperty="common"
       transitionDuration="slow"
-      bg={y > height ? bg : 'transparent'}
       boxShadow={y > height ? 'md' : ''}
-      borderBottomWidth="1px"
+      bg="whiteAlpha.900"
+      _dark={{
+        bg: 'transparent',
+      }}
+      _before={{
+        content: '""',
+        position: 'absolute',
+        inset: 0,
+        bottom: '-30px',
+        backdropFilter: 'blur(16px)',
+        mask: 'linear-gradient(to-b, black, transparent)',
+        '-webkit-mask': 'linear-gradient(to bottom, black 60px, transparent)',
+      }}
       {...props}
     >
-      <Container maxW="container.2xl" px="8" py="4">
-        <HStack width="full" align="center">
-          <Logo
-            onClick={(e) => {
-              if (window.location.pathname === '/') {
-                e.preventDefault()
+      <Box zIndex="1" position="relative" borderBottomWidth="1px">
+        <Container maxW="container.2xl" px="8" py="4">
+          <HStack width="full" align="center">
+            <Logo
+              onClick={(e) => {
+                if (window.location.pathname === '/') {
+                  e.preventDefault()
 
-                window.scrollTo({
-                  top: 0,
-                  behavior: 'smooth',
-                })
-              }
-            }}
-          />
-          <Navigation />
-        </HStack>
-      </Container>
+                  window.scrollTo({
+                    top: 0,
+                    behavior: 'smooth',
+                  })
+                }
+              }}
+            />
+            <Navigation />
+          </HStack>
+        </Container>
+      </Box>
     </Box>
   )
 }