+++ apps/website/src/styles/theme.ts
@@ -1,15 +1,10 @@
-import { extendTheme } from '@chakra-ui/react'
+import { extendTheme, ThemeTypings } from '@chakra-ui/react'
 
 import { theme as baseTheme } from '@saas-ui/pro'
 
 import '@fontsource/inter/variable.css'
 
-import {
-  createBreakpoints,
-  mode,
-  transparentize,
-  blacken,
-} from '@chakra-ui/theme-tools'
+import { mode, transparentize, blacken } from '@chakra-ui/theme-tools'
 
 import Badge from './components/badge'
 import Button from './components/button'
@@ -28,7 +23,7 @@ const styles = {
       webkitFontSmoothing: 'antialiased',
       textRendering: 'optimizeLegibility',
       color: mode('gray.900', 'white')(props),
-      bg: mode('white', 'gray.900')(props),
+      bg: mode('white', 'black')(props),
       minHeight: 'auto',
       fontSize: 'lg',
     },
@@ -71,13 +66,13 @@ const theme = extendTheme(
       initialColorMode: 'dark',
       useSystemColorMode: false,
     },
-    breakpoints: createBreakpoints(breakpoints),
+    breakpoints,
     colors,
     semanticTokens: {
       colors: {
         codeBackground: {
           default: blacken('purple.600', 70)(baseTheme),
-          _dark: 'gray.800',
+          _dark: 'gray.900',
         },
         muted: {
           default: 'gray.600',