+++ packages/saas-ui-sidebar/src/nav-item.tsx
@@ -21,13 +21,18 @@ export interface NavItemLabelProps
   extends HTMLChakraProps<'div'>,
     ThemingProps<'NavLink'> {}
 
-export const NavItemLabel = forwardRef<NavItemLabelProps, 'div'>(
+export const NavItemLabel = forwardRef<NavItemLabelProps, 'span'>(
   ({ children, ...props }, ref) => {
     const styles = useNavItemStyles()
     return (
-      <chakra.div ref={ref} __css={styles.label} {...props}>
+      <chakra.span
+        ref={ref}
+        __css={styles.label}
+        {...props}
+        className={cx('saas-nav-item__label', props.className)}
+      >
         {children}
-      </chakra.div>
+      </chakra.span>
     )
   }
 )
@@ -41,7 +46,7 @@ const NavItemIcon: React.FC<HTMLChakraProps<'span'>> = (props) => {
 
   const child = React.Children.only(children)
 
-  const clone = React.isValidElement(child)
+  const clone = React.isValidElement<HTMLChakraProps<any>>(child)
     ? React.cloneElement(child, {
         focusable: 'false',
         'aria-hidden': true,
@@ -51,6 +56,7 @@ const NavItemIcon: React.FC<HTMLChakraProps<'span'>> = (props) => {
   return (
     <chakra.span
       {...rest}
+      className={cx('saas-nav-item__icon', props.className)}
       __css={{
         flexShrink: 0,
         ...styles.icon,
@@ -67,7 +73,8 @@ export interface NavItemProps
   extends HTMLChakraProps<'a'>,
     ThemingProps<'NavItem'> {
   href?: string
-  label: string
+  /* @deprecated use children instead */
+  label?: string
   icon?: React.ReactElement
   inset?: SystemProps['paddingLeft']
   tooltip?: React.ReactNode
@@ -77,7 +84,7 @@ export interface NavItemProps
 export const NavItem = forwardRef<NavItemProps, 'a'>((props, ref) => {
   const {
     href = '#',
-    label,
+    label: labelProp,
     icon,
     inset,
     className,
@@ -92,12 +99,21 @@ export const NavItem = forwardRef<NavItemProps, 'a'>((props, ref) => {
 
   const styles = useMultiStyleConfig('NavItem', props)
 
+  let label = children || labelProp
+  let tooltipLabel = tooltip
+  if (typeof label === 'string') {
+    if (!tooltipLabel && isCondensed) {
+      tooltipLabel = label
+    }
+    label = <NavItemLabel hidden={isCondensed}>{label}</NavItemLabel>
+  }
+
   let link = (
     <chakra.a
       {...rest}
       ref={ref}
       href={href}
-      className={cx('nav-item', className)}
+      className="nav-item__link"
       data-active={dataAttr(isActive)}
       __css={styles.link}
     >
@@ -106,10 +122,10 @@ export const NavItem = forwardRef<NavItemProps, 'a'>((props, ref) => {
           ...styles.inner,
           pl: inset,
         }}
+        className="saas-nav-item__inner"
       >
         {icon && <NavItemIcon>{icon}</NavItemIcon>}
-        <NavItemLabel>{label}</NavItemLabel>
-        {children}
+        {label}
       </chakra.span>
     </chakra.a>
   )
@@ -118,12 +134,15 @@ export const NavItem = forwardRef<NavItemProps, 'a'>((props, ref) => {
     link = <RouterLink href={href}>{link}</RouterLink>
   }
 
-  const tooltipLabel = isCondensed && !tooltip ? label : tooltip
-
   return (
     <NavItemStylesProvider value={styles}>
       <Tooltip label={tooltipLabel} placement="right" openDelay={400}>
-        <chakra.div __css={styles.item} onClick={onClose}>
+        <chakra.div
+          __css={styles.item}
+          onClick={onClose}
+          data-condensed={dataAttr(isCondensed)}
+          className={cx('saas-nav-item', className)}
+        >
           {link}
         </chakra.div>
       </Tooltip>