+++ packages/saas-ui-sidebar/stories/sidebar.stories.tsx
@@ -3,19 +3,31 @@ import { Story, Meta } from '@storybook/react'
 import {
   Badge,
   BadgeProps,
+  Box,
   Button,
+  DarkMode,
   Flex,
   Heading,
   HStack,
   IconButton,
+  LightMode,
   Menu,
   MenuButton,
   MenuList,
   Spacer,
+  Text,
   useDisclosure,
 } from '@chakra-ui/react'
 
-import { FiHome, FiUsers, FiSettings, FiHash, FiStar } from 'react-icons/fi'
+import {
+  FiHome,
+  FiUsers,
+  FiSettings,
+  FiHash,
+  FiStar,
+  FiChevronsLeft,
+  FiChevronsRight,
+} from 'react-icons/fi'
 
 import { FaHome, FaUsers, FaCog, FaHashtag } from 'react-icons/fa'
 
@@ -159,7 +171,7 @@ WithCollapsibleGroup.args = {
 }
 
 const NavItemBadge = (props: BadgeProps) => (
-  <Badge bg="whiteAlpha.200" fontWeight="normal" rounded="md" {...props} />
+  <Badge bg="none" fontWeight="normal" rounded="md" ms="auto" {...props} />
 )
 
 export const WithBadge = Template.bind({})
@@ -187,14 +199,17 @@ WithBadge.args = {
         </NavGroup>
 
         <NavGroup title="Tags" isCollapsible>
-          <NavItem label="Design system" icon={<FiHash />}>
+          <NavItem icon={<FiHash />}>
+            <Text>Design System</Text>
             <NavItemBadge>1</NavItemBadge>
           </NavItem>
-          <NavItem label="Framework" icon={<FiHash />} />
-          <NavItem label="Chakra UI" inset={5} icon={<FiHash />}>
+          <NavItem icon={<FiHash />}>Framework</NavItem>
+          <NavItem inset={5} icon={<FiHash />}>
+            <Text>Chakra UI</Text>
             <NavItemBadge>32</NavItemBadge>
           </NavItem>
-          <NavItem label="React" inset={5} icon={<FiHash />}>
+          <NavItem inset={5} icon={<FiHash />}>
+            <Text>React</Text>
             <NavItemBadge>100</NavItemBadge>
           </NavItem>
         </NavGroup>
@@ -330,51 +345,90 @@ VariantCondensedColor.args = {
   ),
 }
 
+export const VariantCondensedResponsive = Template.bind({})
+VariantCondensedResponsive.args = {
+  variant: { base: 'condensed', lg: 'default' },
+  breakpoints: { base: false },
+  colorScheme: 'purple',
+  children: (
+    <>
+      <SidebarSection>
+        <Logo width="24px" color="white" />
+      </SidebarSection>
+      <SidebarSection>
+        <NavGroup>
+          <NavItem
+            label="Home"
+            icon={<FiHome size="1.2em" color="whiteAlpha.800" />}
+            color="white"
+            size="sm"
+            isActive
+          />
+          <NavItem
+            label="Users"
+            icon={<FiUsers size="1.2em" color="whiteAlpha.800" />}
+            color="white"
+            size="sm"
+          />
+          <NavItem
+            label="Settings"
+            icon={<FiSettings size="1.2em" color="whiteAlpha.800" />}
+            color="white"
+            size="sm"
+          />
+        </NavGroup>
+      </SidebarSection>
+    </>
+  ),
+}
+
 export const DoubleSidebar = () => {
   const disclosure = useDisclosure()
 
   return (
     <Flex alignItems="stretch" overflow="hidden">
-      <Sidebar
-        variant="condensed"
-        colorScheme="purple"
-        border="0"
-        zIndex="3"
-        position="relative"
-      >
-        <Logo width="24px" color="white" mb="1" />
+      <DarkMode>
+        <Sidebar
+          variant="condensed"
+          colorScheme="purple"
+          border="0"
+          zIndex="3"
+          position="relative"
+        >
+          <Logo width="24px" color="white" mb="1" />
 
-        <NavGroup>
-          <NavItem
-            label="Users"
-            icon={<FiUsers size="1.2em" color="white" />}
-            isActive
-            onClick={(e) => {
-              e.preventDefault()
-              disclosure.onToggle()
-            }}
-          />
-          <NavItem
-            label="Settings"
-            icon={<FiSettings size="1.2em" color="white" />}
-            onClick={(e) => {
-              e.preventDefault()
-              disclosure.onClose()
-            }}
-          />
-        </NavGroup>
+          <NavGroup>
+            <NavItem
+              label="Users"
+              icon={<FiUsers size="1.2em" />}
+              isActive
+              onClick={(e) => {
+                e.preventDefault()
+                disclosure.onToggle()
+              }}
+            />
+            <NavItem
+              label="Settings"
+              icon={<FiSettings size="1.2em" />}
+              onClick={(e) => {
+                e.preventDefault()
+                disclosure.onClose()
+              }}
+            />
+          </NavGroup>
 
-        <Spacer />
+          <Spacer />
 
-        <Menu>
-          <MenuButton as={Button} variant="ghost">
-            <PersonaAvatar presence="online" size="xs" />
-          </MenuButton>
-          <MenuList>
-            <MenuItem>Sign out</MenuItem>
-          </MenuList>
-        </Menu>
-      </Sidebar>
+          <Menu>
+            <MenuButton as={Button} variant="ghost">
+              <PersonaAvatar presence="online" size="xs" />
+            </MenuButton>
+            <MenuList>
+              <MenuItem>Sign out</MenuItem>
+            </MenuList>
+          </Menu>
+        </Sidebar>
+      </DarkMode>
       <Sidebar
         isOpen={disclosure.isOpen}
         onClose={disclosure.onClose}
@@ -404,5 +458,93 @@ export const DoubleSidebar = () => {
         <SidebarOverlay zIndex="1" />
       </Sidebar>
     </Flex>
+  )
+}
+
+export function ToggleSidebar() {
+  const { isOpen, onOpen, onClose, onToggle } = useDisclosure({
+    defaultIsOpen: true,
+  })
+
+  return (
+    <>
+      <Sidebar
+        isOpen={isOpen}
+        onOpen={onOpen}
+        onClose={onClose}
+        breakpoints={{ base: true }}
+      >
+        <Box h="8">
+          <Button onClick={onToggle} position="fixed" left="4" zIndex="modal">
+            {isOpen ? 'Close' : 'Open'}
+          </Button>
+        </Box>
+
+        <SidebarSection ps="6" pe="4" direction="row">
+          <Heading size="sm" py="2">
+            Users
+          </Heading>
+          <Spacer />
+        </SidebarSection>
+
+        <SidebarSection flex="1" overflowY="auto">
+          <NavGroup>
+            <NavItem label="All users" icon={<FiHome />} isActive />
+            <NavItem label="Favourite users" icon={<FiStar />} />
+          </NavGroup>
+          <NavGroup title="Tags" isCollapsible>
+            <NavItem label="Design system" icon={<FiHash />} />
+            <NavItem label="Framework" icon={<FiHash />} />
+            <NavItem label="Chakra UI" inset={5} icon={<FiHash />} />
+            <NavItem label="React" inset={5} icon={<FiHash />} />
+          </NavGroup>
+        </SidebarSection>
+        <SidebarOverlay zIndex="1" />
+      </Sidebar>
+    </>
+  )
+}
+
+export function ToggleVariant() {
+  const { isOpen, onToggle } = useDisclosure({
+    defaultIsOpen: true,
+  })
+
+  return (
+    <>
+      <Sidebar
+        breakpoints={{ base: false }}
+        variant={isOpen ? 'default' : 'condensed'}
+        transition="width"
+        transitionDuration="normal"
+        width={isOpen ? '280px' : '48px'}
+        minWidth="auto"
+      >
+        <SidebarSection direction={isOpen ? 'row' : 'column'}>
+          <Logo
+            width="24px"
+            color="white"
+            mb="1"
+            display={isOpen ? 'block' : 'none'}
+          />
+          <Spacer />
+          <IconButton
+            onClick={onToggle}
+            variant="ghost"
+            size="sm"
+            icon={isOpen ? <FiChevronsLeft /> : <FiChevronsRight />}
+            aria-label="Toggle Sidebar"
+          />
+        </SidebarSection>
+
+        <SidebarSection flex="1" overflowY="auto" overflowX="hidden">
+          <NavGroup>
+            <NavItem label="All users" icon={<FiHome />} isActive />
+            <NavItem label="Favourite users" icon={<FiStar />} />
+          </NavGroup>
+        </SidebarSection>
+        <SidebarOverlay zIndex="1" />
+      </Sidebar>
+    </>
   )
 }