+++ packages/saas-ui-theme/src/base/components/nav-item.ts
@@ -18,6 +18,10 @@ const baseStyle: PartsStyleFunction<typeof parts> = (props) => {
   return {
     item: {
       my: '2px',
+      color: 'gray.900',
+      _dark: {
+        color: 'whiteAlpha.900',
+      },
     },
     link: {
       display: 'flex',
@@ -33,19 +37,20 @@ const baseStyle: PartsStyleFunction<typeof parts> = (props) => {
       _focusVisible: {
         boxShadow: 'outline',
       },
-      '[data-condensed] &': {
+      ['[data-condensed] &']: {
         padding: 0,
-        justifyContent: 'center',
       },
     },
     inner: {
       display: 'flex',
       flex: 1,
       alignItems: 'center',
-      justifyContent: 'center',
+      justifyContent: 'stretch',
+      ['[data-condensed] &']: {
+        justifyContent: 'center',
+      },
     },
     label: {
-      flex: 1,
       '[data-condensed] &': {
         display: 'none',
       },
@@ -64,29 +69,29 @@ const baseStyle: PartsStyleFunction<typeof parts> = (props) => {
 }
 
 const variantNeutral: PartsStyleFunction<typeof parts> = (props) => {
-  const { isActive } = props
-
   const hoverBg = mode('blackAlpha.100', 'whiteAlpha.200')(props)
   const activeBg = mode('blackAlpha.200', 'whiteAlpha.300')(props)
 
   return {
     link: {
-      bg: isActive && activeBg,
       _hover: {
-        bg: isActive ? activeBg : hoverBg,
+        bg: hoverBg,
       },
       _active: {
         bg: activeBg,
       },
     },
     icon: {
-      opacity: isActive ? 1 : 0.8,
+      opacity: 0.8,
+      '[data-active] &': {
+        opacity: 1,
+      },
     },
   }
 }
 
 const variantSubtle: PartsStyleFunction<typeof parts> = (props) => {
-  const { isActive, colorScheme: c, theme } = props
+  const { colorScheme: c, theme } = props
 
   const color = mode(`${c}.700`, `${c}.200`)(props)
   const activeBg = mode(
@@ -94,63 +99,41 @@ const variantSubtle: PartsStyleFunction<typeof parts> = (props) => {
     transparentize(`${c}.500`, 0.3)(theme)
   )(props)
 
-  return isActive
-    ? {
-        link: {
-          bg: activeBg,
-          _hover: {
-            bg: activeBg,
-          },
-          _active: {
-            bg: activeBg,
-          },
-          color,
-        },
-        icon: {
-          color,
-        },
-        label: {
-          fontWeight: 'semibold',
-        },
-      }
-    : {
-        link: {
-          _hover: {
-            bg: mode('blackAlpha.100', 'whiteAlpha.200')(props),
-          },
-        },
-      }
+  return {
+    link: {
+      _hover: {
+        bg: mode('blackAlpha.100', 'whiteAlpha.200')(props),
+      },
+      _active: {
+        bg: activeBg,
+        fontWeight: 'semibold',
+        color: mode(`${c}.600`, `${c}.100`)(props),
+      },
+    },
+  }
 }
 
 const variantSolid: PartsStyleFunction<typeof parts> = (props) => {
-  const { isActive, colorScheme: c } = props
+  const { colorScheme: c } = props
 
   const activeBg = `${c}.500`
+  const hoverBg = mode('blackAlpha.100', 'whiteAlpha.200')(props)
 
-  return isActive
-    ? {
-        link: {
-          bg: activeBg,
-          _hover: {
-            bg: activeBg,
-          },
-          _active: {
-            bg: activeBg,
-          },
-          color: 'white',
-        },
-        icon: {
-          color: 'white',
-        },
-        label: {},
-      }
-    : {
-        link: {
-          _hover: {
-            bg: mode('blackAlpha.100', 'whiteAlpha.200')(props),
-          },
-        },
-      }
+  return {
+    link: {
+      _hover: {
+        bg: hoverBg,
+      },
+      _active: {
+        bg: activeBg,
+      },
+      color: 'white',
+    },
+    icon: {
+      color: 'white',
+    },
+    label: {},
+  }
 }
 
 export default {
@@ -163,28 +146,28 @@ export default {
   baseStyle,
   sizes: {
     xs: {
-      link: baseTheme.components.Button.sizes.xs,
+      link: baseTheme.components.Button.sizes?.xs,
       icon: {
         me: 2,
         fontSize: 'xs',
       },
     },
     sm: {
-      link: baseTheme.components.Button.sizes.sm,
+      link: baseTheme.components.Button.sizes?.sm,
       icon: {
         me: 2,
         fontSize: 'sm',
       },
     },
     md: {
-      link: baseTheme.components.Button.sizes.md,
+      link: baseTheme.components.Button.sizes?.md,
       icon: {
         me: 2,
         fontSize: 'md',
       },
     },
     lg: {
-      link: baseTheme.components.Button.sizes.lg,
+      link: baseTheme.components.Button.sizes?.lg,
       icon: {
         me: 2,
         fontSize: 'lg',