+++ packages/saas-ui-data-table/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @saas-ui/data-table
 
+## 1.3.1
+
+### Patch Changes
+
+- @saas-ui/layout@1.3.1
+
+## 1.3.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [d92f516]
+  - @saas-ui/layout@1.3.0
+
+## 1.2.2
+
+### Patch Changes
+
+- @saas-ui/layout@1.2.2
+
+## 1.2.1
+
+### Patch Changes
+
+- @saas-ui/layout@1.2.1
+
+## 1.2.0
+
+### Minor Changes
+
+- 466b3ca: Updated to Chakra UI 2.3.6
+
+### Patch Changes
+
+- Updated dependencies [466b3ca]
+  - @saas-ui/layout@1.2.0
+
 ## 1.1.2
 
 ### Patch Changes