+++ packages/saas-ui-forms/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @saas-ui/forms
 
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies [7cf5223]
+  - @saas-ui/react-utils@1.1.2
+  - @saas-ui/stepper@1.3.1
+
+## 1.4.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [d92f516]
+  - @saas-ui/button@1.3.0
+  - @saas-ui/input-right-button@1.3.0
+  - @saas-ui/number-input@1.3.0
+  - @saas-ui/password-input@1.3.0
+  - @saas-ui/pin-input@1.3.0
+  - @saas-ui/radio@1.3.0
+  - @saas-ui/select@1.3.0
+  - @saas-ui/stepper@1.3.0
+
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [92f3e73]
+  - @saas-ui/react-utils@1.1.1
+  - @saas-ui/stepper@1.2.1
+
+## 1.3.0
+
+### Minor Changes
+
+- 466b3ca: Updated to Chakra UI 2.3.6
+
+### Patch Changes
+
+- Updated dependencies [466b3ca]
+  - @saas-ui/button@1.2.0
+  - @saas-ui/input-right-button@1.2.0
+  - @saas-ui/number-input@1.2.0
+  - @saas-ui/password-input@1.2.0
+  - @saas-ui/pin-input@1.2.0
+  - @saas-ui/radio@1.2.0
+  - @saas-ui/select@1.2.0
+  - @saas-ui/stepper@1.2.0
+
 ## 1.2.2
 
 ### Patch Changes