+++ packages/saas-ui-palette/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.2.0
 
+## 1.2.0
+
+### Minor Changes
+
+- 466b3ca: Updated to Chakra UI 2.3.6
+
 ## 1.1.0
 
 ### Minor Changes