+++ packages/saas-ui-property/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @saas-ui/property
 
+## 1.3.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [d92f516]
+  - @saas-ui/list@1.3.0
+
+## 1.2.0
+
+### Minor Changes
+
+- 466b3ca: Updated to Chakra UI 2.3.6
+
+### Patch Changes
+
+- Updated dependencies [466b3ca]
+  - @saas-ui/list@1.2.0
+
 ## 1.1.0
 
 ### Minor Changes