+++ packages/saas-ui-react-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @saas-ui/react-utils
 
+## 1.1.2
+
+### Patch Changes
+
+- 7cf5223: useResponsive value now uses correct fallback value.
+
+## 1.1.1
+
+### Patch Changes
+
+- 92f3e73: Sidebar now supports responsive variant value
+
 ## 1.1.0
 
 ### Minor Changes