+++ tooling/props-docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @saas-ui/props-docs
 
+## 1.7.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
+## 1.6.0
+
+### Minor Changes
+
+- 466b3ca: Updated to Chakra UI 2.3.6
+
 ## 1.5.0
 
 ### Minor Changes